refactor(ui): extract terminal lifecycle to ui/terminal.rs
setup_terminal, restore_terminal, and the Terminal<B> type alias
move into their own module. Re-exported from ui::* so main.rs sees
the same ui::setup_terminal API.
