+++ src/ui/mod.rs
@@ -8,7 +8,7 @@ use std::sync::atomic::{AtomicBool, Ordering};
 
 use anyhow::Result;
 use ratatui::{
-    Frame, Terminal as RatatuiTerminal,
+    Frame,
     layout::{Constraint, Direction, Layout, Rect},
     style::{Color, Modifier, Style},
     symbols,
@@ -25,7 +25,8 @@ use crate::network::types::{
     AppProtocolDistribution, Connection, Protocol, ProtocolState, TcpState, TrafficHistory,
 };
 
-pub type Terminal<B> = RatatuiTerminal<B>;
+mod terminal;
+pub use terminal::{Terminal, restore_terminal, setup_terminal};
 
 /// Placeholder string displayed when a value is unavailable.
 const NONE_PLACEHOLDER: &str = "-";
@@ -213,38 +214,6 @@ pub enum GroupedRow<'a> {
         connection: &'a Connection,
         is_last_in_group: bool,
     },
-}
-
-/// Set up the terminal for the TUI application
-pub fn setup_terminal<B: ratatui::backend::Backend>(backend: B) -> Result<Terminal<B>>
-where
-    <B as ratatui::backend::Backend>::Error: Send + Sync + 'static,
-{
-    let mut terminal = RatatuiTerminal::new(backend)?;
-    terminal.clear()?;
-    terminal.hide_cursor()?;
-    crossterm::terminal::enable_raw_mode()?;
-    crossterm::execute!(
-        std::io::stdout(),
-        crossterm::terminal::EnterAlternateScreen,
-        crossterm::event::EnableMouseCapture
-    )?;
-    Ok(terminal)
-}
-
-/// Restore the terminal to its original state
-pub fn restore_terminal<B: ratatui::backend::Backend>(terminal: &mut Terminal<B>) -> Result<()>
-where
-    <B as ratatui::backend::Backend>::Error: Send + Sync + 'static,
-{
-    crossterm::terminal::disable_raw_mode()?;
-    crossterm::execute!(
-        std::io::stdout(),
-        crossterm::terminal::LeaveAlternateScreen,
-        crossterm::event::DisableMouseCapture
-    )?;
-    terminal.show_cursor()?;
-    Ok(())
 }
 
 /// Represents an action that can be triggered by clicking a screen region.
